refactor(frontend): standardize filenames to kebab-case
Rename App.tsx, sourcePanel.ts, the useXxx hooks, and the __test__
fixtures directory so every frontend file follows the kebab-case
convention already used by the rest of the tree. Updates all imports
(including the e2e spec that pulls from `src/sourcePanel`) accordingly.
