+++ internal/frontend/src/components/file-tree/tree.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 import type { FileEntry } from "../../api/files";
-import { flatFiles, nestedFiles } from "./__test__/fixtures";
+import { flatFiles, nestedFiles } from "./test/fixtures";
 import { buildTree, flatten } from "./tree";
 
 describe("buildTree", () => {
