+++ internal/frontend/src/hooks/use-active-render.ts
