+++ internal/frontend/src/hooks/use-server-events.ts
