+++ internal/frontend/src/source-panel.ts
